handle stereo speech properly
